requirements.txt
Django==4.1.3
djangorestframework~=3.14.0
psycopg2==2.9.6
django-environ==0.10.0